Reverted change in the Efficiency calculation
The Efficiency calculation is fairly low level. It was misguided to let it
operate on estimations (which are a subclass of HigherOrderCalculation).
Efficiencies for a complete background model must be done at a higher level,
and these are complicated and probably unnecessary to cache.

Factored out the actual efficiency calculation and placed it in
owls_hep.utility, to make it accessible from plotting scripts.