Plotting improvements
Added a generic utility.clone function that also handles TF1 objects.
Added plotting geometry for the stat box with a ratio plot.
Adjusted properties for the ratio plot.
Improved and expanded treatment of ratio plots to also handle multiple
histograms and not just data/model.
Added drawing of functions (TF1).
Added an optional parameter to draw_legend to place functions in the legend.