Added more utility functions
Added load_file which loads a ROOT file in a specific mode.
Added helper function to get bins and errors from an object irregardles of
whether it's a TGraph or a TH1.
Added helper function to print bin centers, bin contents, and errors for a
TGraph or a TH1.
Turn on Sumw2() for all created THn's.