Fixed two bugs pertaining to histograms
Reset overflow bin after adding it to the last bin (to avoid double counting
if function is called again)
Set bin content of total bin to 0.0 instead of 1.0, if content <= 0. Setting
to 1.0 conflicts with adding overflow bin to histogram. (But perhaps
the correct fix that _make_consistent() should not act on underflow and
overflow bins.)